api-medicos/requirements.txt
fastapi[all]
redis
uvicorn
python-dotenv
